requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
flask==3.0.2
yfinance==0.2.38
alpha_vantage==2.3.1
